// svelte/src/generated/hello-world/models/HelloWorldSearchQueryResponseMessagePagination.ts
// tslint:disable
/**
 * Hello World API
 * A \"Hello World!\" API
 *
 * The version of the OpenAPI document: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

/**
 * Pagination data
 * @export
 * @interface HelloWorldSearchQueryResponseMessagePagination
 */
export interface HelloWorldSearchQueryResponseMessagePagination {
    /**
     * The page that was queried.
     * @type {number}
     * @memberof HelloWorldSearchQueryResponseMessagePagination
     */
    page: number;
    /**
     * The number of items included in each page.
     * @type {number}
     * @memberof HelloWorldSearchQueryResponseMessagePagination
     */
    pageSize: number;
    /**
     * The total number of items that match the query criteria.
     * @type {number}
     * @memberof HelloWorldSearchQueryResponseMessagePagination
     */
    totalResults: number;
}